Allow client to specify elapsed_time on pause.
This at reduces latency issues when the marker presses the pause button -- all other clients may take a while before they stop their clocks, but when they do stop their clocks they will all show the correct time.

This is also important to prevent the clock from drifting while multiple penalties are being marked.

+++ app/models/timer.rb
@@ -76,12 +76,12 @@ class Timer < ActiveRecord::Base
 
   def elapsed_time=(val)
     if val.is_a?(String)
-      val = val.split(":").map(&:to_i)
+      val = val.split(":").map(&:to_f)
       seconds = val.pop
       seconds += val.pop * 60 if val.any?
       seconds += val.pop * 3600 if val.any?
     else
-      seconds = val.to_i
+      seconds = val.to_f
     end
     self.seconds_paused = diff_in_seconds(paused? ? paused_at : DateTime.now, started_at) - seconds
   end
